requirements.txt
altair==5.4.1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
catboost==1.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
executing==2.0.1
filelock==3.16.0
fonttools==4.49.0
fsspec==2024.9.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.2
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
narwhals==1.9.3
nest-asyncio==1.6.0
networkx==3.4
numpy==1.26.4
opencv-python==4.9.0.80
openvino==2024.4.1
openvino-telemetry==2024.1.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==5.28.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycairo==1.27.0
pydeck==0.9.1
Pygments==2.17.2
PyGObject==3.50.0
pyparsing==3.1.2
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.35.1
requests==2.31.0
rich==13.9.2
rpds-py==0.20.0
scipy==1.12.0
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.3
tenacity==8.2.3
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.9
ultralytics-thop==2.0.9
urllib3==2.2.1
videosplitter @ file:///Users/mike-yasnov/Desktop/Work/videos_splitter
wcwidth==0.2.13
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.7_1/libexec/wheel-0.44.0-py3-none-any.whl#sha256=c01326144883c50263d6967d1145640f3bbc8dfab32c2cbe817ef61d7a810219
